Trying to get the wfc to work with async

+++ The Game/Assets/Scripts/WFC/CoreSolver/PropagationHelper.cs
@@ -50,7 +50,7 @@ namespace WaveFunctionCollaps
 
             if (elementIdLowEntropySet == null && outputGrid.CheckIfCellIsCollapsed(cellToPropagatePosition) == false)
             {
-                float entropy = coreHelper.CalculateEntropy(cellToPropagatePosition, outputGrid);
+                double entropy = coreHelper.CalculateEntropy(cellToPropagatePosition, outputGrid);
                 lowestEntropySet.Add(new LowEntropyCell(cellToPropagatePosition, entropy));
 
             }
